+++ packages/polaris-viz-core/src/types.ts
@@ -27,6 +27,9 @@ interface StyleOverride {
     width?: number;
     strokeDasharray?: string;
   };
+  tooltip?: {
+    shape?: Shape;
+  };
 }
 
 export interface DataGroup {