+++ packages/polaris-viz/src/components/LineChartRelational/components/index.ts
@@ -1,3 +1,4 @@
 export {Area} from './Area';
 export {MissingDataArea} from './MissingDataArea';
 export {RelatedAreas} from './RelatedAreas';
+export {CustomLegend} from './CustomLegend';